requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.1.0

# Graph Database
neo4j>=5.0.0
py2neo>=2021.2.3

# AI/ML libraries
google-genai>=0.8.0
scikit-learn>=1.1.0
torch>=2.1.0
transformers>=4.35.0

# Graph Analytics
networkx>=2.8
python-louvain>=0.16
scipy>=1.11.0
cdlib>=0.2.6
node2vec>=0.4.6

# Visualization
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyvis>=0.3.2
dash>=2.15.0

# Document processing
docx2txt>=0.8
python-docx>=0.8.11
pypandoc>=1.5

# Web App Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-plotly-events>=0.0.6

# Utilities
tqdm>=4.64.0
loguru>=0.6.0
rich>=12.0.0
python-dotenv>=1.0.0
click>=8.1.0
jsonlines>=3.1.0
joblib>=1.3.0
requests>=2.31.0

# Configuration
toml>=0.10.2
pydantic>=2.5.0
pyyaml>=6.0

# Advanced Analytics
scipy>=1.11.0
cdlib>=0.2.6
node2vec>=0.4.6
statsmodels>=0.14.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
black>=22.0.0

# Analysis
jupyter>=1.0.0
ipywidgets>=8.1.0
statsmodels>=0.14.0 